+++ src/main/java/racingcar/domain/car/CarName.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 public record CarName(String name) {
