+++ src/main/java/racingcar/domain/car/strategy/CarMovementStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain.strategy;
+package racingcar.domain.car.strategy;
 
 @FunctionalInterface
 public interface CarMovementStrategy {
