+++ src/main/java/racingcar/domain/car/strategy/MoveForwardOnceRandomlyStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain.strategy;
+package racingcar.domain.car.strategy;
 
 import camp.nextstep.edu.missionutils.Randoms;
