+++ src/main/java/racingcar/domain/messenger/Messenger.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.messenger;
 
 import racingcar.dto.CarNamesDto;
 import racingcar.dto.CarsInformationDto;
