+++ src/test/java/racingcar/domain/car/CarEqualsTest.java
@@ -1,12 +1,10 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.CarName;
-import racingcar.domain.strategy.CarMovementStrategy;
-import racingcar.domain.strategy.MoveForwardOnceRandomlyStrategy;
+import racingcar.domain.car.strategy.CarMovementStrategy;
+import racingcar.domain.car.strategy.MoveForwardOnceRandomlyStrategy;
 
 public class CarEqualsTest {
