+++ src/test/java/racingcar/domain/car/CarGetInformationTest.java
@@ -1,10 +1,9 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.strategy.MoveForwardOnceRandomlyStrategy;
+import racingcar.domain.car.strategy.MoveForwardOnceRandomlyStrategy;
 import racingcar.dto.CarInformationDto;
 
 public class CarGetInformationTest {
