+++ src/test/java/racingcar/domain/car/CarNameLengthValidateTest.java
@@ -1,11 +1,10 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.domain.CarName;
 
 public class CarNameLengthValidateTest {
