+++ src/test/java/racingcar/domain/car/CarPositionMoveTest.java
@@ -1,11 +1,10 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.domain.CarPosition;
 
 public class CarPositionMoveTest {
