+++ src/test/java/racingcar/domain/car/CarsGeneratorTest.java
@@ -1,14 +1,11 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
-import racingcar.domain.Cars;
-import racingcar.domain.CarsGenerator;
-import racingcar.domain.strategy.CarMovementStrategy;
-import racingcar.domain.strategy.MoveForwardOnceRandomlyStrategy;
+import racingcar.domain.car.strategy.CarMovementStrategy;
+import racingcar.domain.car.strategy.MoveForwardOnceRandomlyStrategy;
 
 public class CarsGeneratorTest {
