+++ src/test/java/racingcar/domain/round/RoundValidateIsPositiveTest.java
@@ -1,11 +1,10 @@
-package racingcar.round;
+package racingcar.domain.round;
 
-        import org.junit.jupiter.api.Assertions;
-        import org.junit.jupiter.api.DisplayName;
-        import org.junit.jupiter.api.Test;
-        import org.junit.jupiter.params.ParameterizedTest;
-        import org.junit.jupiter.params.provider.CsvSource;
-        import racingcar.domain.Round;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 public class RoundValidateIsPositiveTest {
